+++ ChangeLog.md
@@ -14,6 +14,7 @@
 - Windows/ControlsConsole use a theme for drawing.
 - InputBox renamed TextBox.
 
+
 ## 06/11/2018
 
 - Fix a stack overflow problem in the window object introduced by the previous mouse bug fix.